fix broken status bar and table switching
for some reason the PackageTableModel class requires the parent to be passed to it and set as an attribute in order for the combobox in the parent that allows switching between the tables to work properly. Even though no errors are raised and the attribute isn't referenced anywhere in the class, the combobox's: `combo_version.activated.connect(self.switch_package_table)` fails to trigger without it. Something to do with how Qt/PySide6 works under the hood.