naive & smart (broken) way to get hconfig env vars
this does not actually fix the issue of HPM not being able to recognize newer Houdini versions due to their hconfig's being incompatible with HPM's python version. Due to a hard to fix bug, the smart way of running hconfig with a compatible python version has been provided even though it still fails to return the correct data. If this is fixed in the future it will be the ideal solution to always correctly interface with any hconfig version.