fetch latest tags for pkgs
Includes buttons per package row that fetch all tag versions from the pkg's remote repo via the GitHub API. This new tag data then updates the row in the package table.

Currently the fetched data is not written to disk and is therefore lost when HPM is restarted.

A lot of changes and additions to the model and view backend functionality had to be made in this commit to make syncing git tags possible. That's why the commit is so big.